Address additional code review feedback
- Remove HTML markdownlint disable/enable comments in rust-doctest-dry-guide.md
  and reflow paragraph to wrap at 80 columns
- Update frontmatter documentation to clarify that YAML closers accept trailing
  whitespace (matching is done after trim_end())
- Remove duplicated Rustdoc for process_stream_opts in src/process.rs
- Extract frontmatter CLI tests from tests/cli.rs to new tests/cli_frontmatter.rs
  to bring cli.rs under 400 lines (now 392 lines)
